refactor: change order id type from Long to Integer to align with database schema

+++ backend/src/main/java/com/nhom5/backend/controller/OrderController.java
@@ -58,7 +58,7 @@ public class OrderController {
     }
 
     @PutMapping("/{orderId}/cancel")
-    public ResponseEntity<?> cancelOrder(@PathVariable Long orderId, @RequestParam Integer userId) {
+    public ResponseEntity<?> cancelOrder(@PathVariable Integer orderId, @RequestParam Integer userId) {
         try {
             OrderResponse response = orderService.cancelOrder(orderId, userId);
             return ResponseEntity.ok(response);
